pyproject.toml
[project]
name = "ufload3"
dynamic = ["version"]
authors = [
  { name="MSF" },
]
description = "UniField  python3 loader"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Office365-REST-Python-Client==2.6.2",
]

[project.urls]
Homepage = "https://github.com/Unifield/ufload3"

[project.scripts]
ufload3 = "ufload3.cli.main:main"

[tool.setuptools.dynamic]
version = {attr = "ufload3.__version__"}
